common component for Notification Banners
