+++ ui/src/app/page/signed-in/login-gov-ial2-notification.spec.tsx
@@ -1,13 +1,13 @@
 import {mount} from 'enzyme';
 import * as React from 'react';
 import * as fp from 'lodash/fp';
+import {MemoryRouter} from 'react-router-dom';
 
 import {LoginGovIAL2NotificationMaybe} from "./login-gov-ial2-notification";
 import {profileStore, serverConfigStore} from 'app/utils/stores';
 import defaultServerConfig from 'testing/default-server-config';
 import {AccessModule} from 'generated/fetch';
 import {ProfileStubVariables} from 'testing/stubs/profile-api-stub';
-import { MemoryRouter } from 'react-router-dom';
 
 describe('LoginGovIAL2Notification', () => {
   const load = jest.fn();
@@ -30,19 +30,19 @@ describe('LoginGovIAL2Notification', () => {
         enableRasLoginGovLinking: true
       }});
     const wrapper = component();
-    expect(wrapper.find('[data-test-id="ial2-notification"]').length).toBe(2);
+    expect(wrapper.find('[data-test-id="ial2-notification"]').exists()).toBeTruthy();
   });
 
-  it('should not notification when login.gov feature flag disabled', () => {
+  it('should not show notification when login.gov feature flag disabled', () => {
     serverConfigStore.set({config: {
         ...defaultServerConfig,
         enableRasLoginGovLinking: false
       }});
     const wrapper = component();
-    expect(wrapper.find('[data-test-id="ial2-notification"]').length).toBe(0);
+    expect(wrapper.find('[data-test-id="ial2-notification"]').exists()).toBeFalsy();
   });
 
-  it('should not notification when login.gov complete', () => {
+  it('should not show notification when login.gov complete', () => {
     serverConfigStore.set({config: {
         ...defaultServerConfig,
         enableRasLoginGovLinking: true
@@ -53,10 +53,10 @@ describe('LoginGovIAL2Notification', () => {
     profileStore.set({profile: newProfile, load, reload, updateCache});
 
     const wrapper = component();
-    expect(wrapper.find('[data-test-id="ial2-notification"]').length).toBe(0);
+    expect(wrapper.find('[data-test-id="ial2-notification"]').exists()).toBeFalsy();
   });
 
-  it('should not notification when login.gov bypassed', () => {
+  it('should not show notification when login.gov bypassed', () => {
     serverConfigStore.set({config: {
         ...defaultServerConfig,
         enableRasLoginGovLinking: true
@@ -67,7 +67,7 @@ describe('LoginGovIAL2Notification', () => {
     profileStore.set({profile: newProfile, load, reload, updateCache});
 
     const wrapper = component();
-    expect(wrapper.find('[data-test-id="ial2-notification"]').length).toBe(0);
+    expect(wrapper.find('[data-test-id="ial2-notification"]').exists()).toBeFalsy();
   });
 
 });
